+++ src/components/modals/DeleteLinkModal.tsx
@@ -5,7 +5,7 @@ import { useFocusTrap } from "@/src/hooks/useFocusTrap.ts";
 import { focusRingClasses } from "@/src/styles/focusRing.ts";
 import { getSpriteUrlForPokemonName } from "@/src/services/sprites.ts";
 import { FiAlertTriangle, FiInfo } from "react-icons/fi";
-import Tooltip from "@/src/components/basic/Tooltip.tsx";
+import Tooltip from "@/src/components/other/Tooltip.tsx";
 
 interface DeleteLinkModalProps {
   isOpen: boolean;
